+++ src/main/java/de/maxhenkel/resourcepackchecker/IFilePackResource.java
@@ -1,9 +1,0 @@
-package de.maxhenkel.resourcepackchecker;
-
-import java.io.File;
-
-public interface IFilePackResource {
-
-    File getFile();
-
-}