+++ src/main/java/de/maxhenkel/resourcepackchecker/interfaces/ClearableServerPackManager.java
@@ -1,0 +1,7 @@
+package de.maxhenkel.resourcepackchecker.interfaces;
+
+public interface ClearableServerPackManager {
+
+    void clear();
+
+}