+++ src/main/java/de/maxhenkel/resourcepackchecker/interfaces/FilePackResource.java
@@ -1,0 +1,11 @@
+package de.maxhenkel.resourcepackchecker.interfaces;
+
+import javax.annotation.Nullable;
+import java.io.File;
+
+public interface FilePackResource {
+
+    @Nullable
+    File getFile();
+
+}