+++ src/main/java/de/maxhenkel/resourcepackchecker/interfaces/PackRepository.java
@@ -1,0 +1,9 @@
+package de.maxhenkel.resourcepackchecker.interfaces;
+
+import net.minecraft.server.packs.repository.RepositorySource;
+
+public interface PackRepository {
+
+    void addSource(RepositorySource source);
+
+}