+++ src/main/java/de/maxhenkel/resourcepackchecker/mixin/FilePackResourcesMixin.java
@@ -1,6 +1,6 @@
 package de.maxhenkel.resourcepackchecker.mixin;
 
-import de.maxhenkel.resourcepackchecker.IFilePackResource;
+import de.maxhenkel.resourcepackchecker.interfaces.FilePackResource;
 import net.minecraft.server.packs.FilePackResources;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import java.io.File;
 
 @Mixin(FilePackResources.class)
-public class FilePackResourcesMixin implements IFilePackResource {
+public class FilePackResourcesMixin implements FilePackResource {
 
     @Shadow
     @Final