+++ changelog.md
@@ -1,1 +1,3 @@
-- Updated dependencies
+- Fixed not being able to join servers with a server resource pack
+- Added list_cached_server_packs config option
+- Added keep_cached_server_packs config option